feat: load ships
There are some logic bugs present in this implementation.
The max value of the range input should be the amount in the warehouse
plus ship cargo of that ware. But then I was able to have negative
amount in the warehouse.
If I dynamically adjust the max value I struggle with unload.

I expect this to be fixed once I implement unload of ships.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>