feat: accept playername input
I'm not handling the information right now.
It must be entered before I allow to move to the next scene. This
validation logic is not present. I should also strip „dangerous”
characters that could be interpreted as code or markup before committing
the value to the store. Therefore I will display the „sanitized”
information as feedback.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>