feat: start with implenting transits
I started this commit a week ago but could not complete it because a
refactoring was required prior.

It implements an event delegation, the action and an uncomplete reducer.
I need to check the month of departure against the active month while
considering the distance. It should be easier to do ao on a laptop.
Types and tests are missing.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>