refactor: reorder and rename
Since I resolve CSS @import in order I figured I should use an index
file to ensure the „correct” one. This might go into the 7-in-1
direction of SASS, let's see.

I also updated the rollup config to generate an app.js, because there is
no server component to this game. I also used the chance to set the game
name in the <title>.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>