feat: add checkOnWinConditionReducer
This is my first approach: wait long enough.

I plan to introduce a currency (silver coin?) eventually together with a
little story. It goes as this: At the beginning of your adventure some
fellow merchants invests in you and expects a return a year and a month
(= 13 months) later. You win when you have made more money than with
what you started with, otherwise you loose.

For the free play I want to have the end of the game once you enter a
new century. I have not spend time thinking about how I want to evaluate
the win here.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>