feat: dispatch setTutorialAction on scene nav
I now dispatch a second action whenever I navigate to another scene.
This allows for setting the tutorial flags. I must test whether it
causes the view to flicker in which case I would need to swap the order
of the actions or move the dispatch to the onInput event listener.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>