refactor: introduce warehouse.level
In order to impose a limit on the amount of goods one can hold in a
warehouse I add the property of level. This required me to refactor the
existing shape of the warehouse property to move the goods into a stock
property. It introduced several adjustments in the tests and
implementation.

This way I can introduce events like the burn down of warehouses
(resetting the level) that introduced the popularity of the bricks the
North of Germany is now known for.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>