refactor: remove Stralsund
I have to save some bytes and the cities state is responsible for the
majority of the bundle. Therefore I researched which city were known for
what and decided that I can let go of Stralsund. Its importance only
rose in the 14th century with respect to a peace with Denmark.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>